optimize/requirements.txt
google-ads
google-cloud
google-cloud-bigquery
grpcio
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
pandas
Flask